requirements.txt
pydantic==1.9.1
toml==0.10.2
redis==4.5.4
orjson==3.8.10
loguru==0.6.0
tortoise-orm[asyncmy]==0.19.2
aerich==0.6.3
uvicorn[standard]==0.18.2
asyncssh==2.11.0
aiomysql==0.1.1
fastapi==0.79.0
bcrypt==4.0.1
async_property==0.2.1
werkzeug==2.3.0
arq==0.25.0
fastapi-crudrouter==0.8.6
PyJWT==2.6.0
gunicorn==20.1.0